FFT moving update

+++ STM32/Src/fft.h
@@ -14,8 +14,8 @@
 #define FFT_TARGET 6.0f                                          // average threshold of the FFT signal
 #define FFT_MAX 8.0f                                             // MAX FFT signal threshold
 #define FFT_STEP_COEFF 10.0f                                     // step coefficient for auto-calibration of the FFT signal (more - slower)
-#define FFT_HZ_IN_PIXEL (IQ_SAMPLERATE / LAY_FFT_PRINT_SIZE)     // hertz per pixel
-#define FFT_TX_HZ_IN_PIXEL (TRX_SAMPLERATE / LAY_FFT_PRINT_SIZE) // hertz per pixel
+#define FFT_HZ_IN_PIXEL (float32_t)((float32_t)IQ_SAMPLERATE / (float32_t)LAY_FFT_PRINT_SIZE)     // hertz per pixel
+#define FFT_TX_HZ_IN_PIXEL (float32_t)((float32_t)TRX_SAMPLERATE / (float32_t)LAY_FFT_PRINT_SIZE) // hertz per pixel
 #define FFT_BW_BRIGHTNESS 5																		 // pixel brightness on bw bar
 
 // Public variables
